rf: SetupReplication class in separate file
